+++ src/main/frontend/src/store/sagas/employeesSaga.js
@@ -1,15 +1,10 @@
 import { call, put, takeLatest } from 'redux-saga/effects';
 import {
-  clearEmployees,
   errorReceivingEmployees,
   receiveEmployees,
-  requestEmployees,
-  searchEmployees
+  requestEmployees
 } from '../actions/employeesActions';
-import {
-  getEmployeesApi,
-  searchEmployeesApi
-} from '../services/employeesService';
+import { getEmployeesApi } from '../services/employeesService';
 
 export const getEmployees = function*() {
   try {
@@ -20,23 +15,4 @@ export const getEmployees = function*() {
   }
 };
 
-export const searchEmployeesSaga = function*(action) {
-  if (action.payload) {
-    try {
-      const employeesInformation = yield call(
-        searchEmployeesApi,
-        action.payload
-      );
-      yield put(receiveEmployees(employeesInformation));
-    } catch (error) {
-      yield put(errorReceivingEmployees(error));
-    }
-  } else {
-    yield put(clearEmployees());
-  }
-};
-
-export const employeesSaga = [
-  takeLatest(requestEmployees, getEmployees),
-  takeLatest(searchEmployees, searchEmployeesSaga)
-];
+export const employeesSaga = [takeLatest(requestEmployees, getEmployees)];
